slight change in BlockMaker, Inverse and Product

+++ Armstrong Numbers/Crypto/src/crypto/Inverse.java
@@ -10,9 +10,9 @@ package crypto;
  * @author abhil
  */
 public class Inverse {
-    public double[][] getInverse(double mat[][])
+    public float[][] getInverse(float mat[][])
     {
-        double invmat[][] = new double[mat.length][mat[0].length];
+        float invmat[][] = new float[mat.length][mat[0].length];
         int cofact[][]= new int[mat.length][mat[0].length];
         try{
         // Product of Minors
@@ -58,7 +58,7 @@ public class Inverse {
             }
         }
         
-        double determinant = (mat[0][0] * ((mat[1][1] * mat[2][2]) - (mat[1][2] * mat[2][1])) - mat[0][1] * ((mat[1][0] * mat[2][2]) - (mat[1][2] * mat[2][0])) + mat[0][2] * ((mat[1][0] * mat[2][1]) - (mat[1][1] * mat[2][0])));
+        float determinant = (mat[0][0] * ((mat[1][1] * mat[2][2]) - (mat[1][2] * mat[2][1])) - mat[0][1] * ((mat[1][0] * mat[2][2]) - (mat[1][2] * mat[2][0])) + mat[0][2] * ((mat[1][0] * mat[2][1]) - (mat[1][1] * mat[2][0])));
         
         // Inverse
         for(int i=0; i<invmat.length; i++)
